Merging SoilMoisturePoller and WateringEventPoller
It turns out things are a lot simpler when we combine the poller for checking
soil moisture with the poller responsible for running the pump based on soil
moisture. Now, the SoilWateringPoller can check soil moisture, store the
record in the DB, but then use the same reading to determine if it needs to
pump water.

Because of this, we can get rid of the get_latest method on SoilMoistureStore,
which is nice because now all the DbStore classes have the exact sameinterface
and we can delete the old SoilMoisturePoller and not add much code to the
former WateringEventPoller.