Changing polling to occur on time boundaries
In the current implementation, each sensor starts reading immediately, then
polls every n seconds. This means that the readings depend on when the
greenpithumb started, so if a user runs greenpithumb on two different days
with two different start times, the reading timestamps won't be aligned.

By changing the polling time to happen on interval boundaries, we get aligned
readings no matter when we start the greenpithumb process.

As a side effect, this doesn't solve the uninterruptible process issue (#99)
but it should make it easier to fix since we're getting rid of most of the
threads that block forever until some event happens.