Changing poll_interval to be a timedelta
This fixes a bug in poller where it was treating the interval as seconds in
one spot and minutes in another.

This is really a temporary workaround until we refactor out the poll
scheduling logic because I think this bug shows we're not testing it
sufficiently and the way the code is currently organized, it's hard to test
the poll scheduling part.