Switching from PlantFriends soil sensor to SparkFun
Changes code to use the SparkFun soil moisture sensor instead of the
PlantFriends soil moisture sensor. The SparkFun cannot reverse its polarity
like the PlantFriends, so it is constantly connected to ground and only needs
one GPIO pin to flip up and down for readings.

Updates the code and the Fritzing diagram.